Fix compilation error

+++ src/main/java/com/youcruit/billogram/objects/response/item/Item.java
@@ -43,7 +43,7 @@ public class Item extends BillogramItem {
     }
 
     @Override
-    public Item withDiscount(Integer discount) {
+    public Item withDiscount(BigDecimal discount) {
         return (Item) super.withDiscount(discount);
     }
